corrected structure
