+++ src/core/Problems/Gaussian2DPulse.cpp
@@ -15,33 +15,27 @@ namespace WaveEquation
             return std::exp(-r_squared / (2.0 * sigma * sigma));
         }
         
-        virtual double initial_velocity(const dealii::Point<dim> &p) const override
+        virtual double initial_velocity(const dealii::Point<dim> &/*p*/) const override
         {
             // Zero initial velocity for symmetric expansion
-            (void)p;  // Suppress unused parameter warning
             return 0.0;
         }
         
-        virtual double boundary_value(const dealii::Point<dim> &p, double t) const override
+        virtual double boundary_value(const dealii::Point<dim> &/*p*/, double /*t*/) const override
         {
             // Homogeneous Dirichlet boundary conditions
-            (void)p;
-            (void)t;
             return 0.0;
         }
         
-        virtual double source_term(const dealii::Point<dim> &p, double t) const override
+        virtual double source_term(const dealii::Point<dim> &/*p*/, double /*t*/) const override
         {
             // No external forcing
-            (void)p;
-            (void)t;
             return 0.0;
         }
         
-        virtual double value(const dealii::Point<dim> &p, const unsigned int component = 0) const override
+        virtual double value(const dealii::Point<dim> &p, const unsigned int /*component*/) const override
         {
             // For dealii::Function interface - used for boundary conditions
-            (void)component;
             return boundary_value(p, this->get_time());
         }
     };
