+++ src/core/Utilities/EnergyCalculator.cpp
